fix: address endpoint discovery review nits
