+++ src/core/config/types/config_impls.rs
@@ -277,6 +277,16 @@ impl fmt::Debug for Config {
             .field("discover_sitemaps", &self.discover_sitemaps)
             .field("sitemap_since_days", &self.sitemap_since_days)
             .field("map_fallback", &self.map_fallback)
+            .field("endpoints_include_bundles", &self.endpoints_include_bundles)
+            .field(
+                "endpoints_first_party_only",
+                &self.endpoints_first_party_only,
+            )
+            .field("endpoints_unique_only", &self.endpoints_unique_only)
+            .field("endpoints_max_scripts", &self.endpoints_max_scripts)
+            .field("endpoints_max_scan_bytes", &self.endpoints_max_scan_bytes)
+            .field("endpoints_verify", &self.endpoints_verify)
+            .field("endpoints_capture_network", &self.endpoints_capture_network)
             .field("max_sitemaps", &self.max_sitemaps)
             .field("cache", &self.cache)
             .field("cache_http_only", &self.cache_http_only)
